# [WIP] Vulkan: use 8/16 bit writes when possible