--- a/Source/gs/GSH_Vulkan/GSH_VulkanMemoryUtils.cpp
+++ b/Source/gs/GSH_Vulkan/GSH_VulkanMemoryUtils.cpp
@@ -64,6 +64,13 @@ void CMemoryUtils::Memory_Write24(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayU
 	AtomicOr(memoryBuffer, wordAddress, value);
 }
 
+void CMemoryUtils::Memory_Write24(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUint8Value memoryBuffer, Nuanceur::CIntValue address, Nuanceur::CUintValue value)
+{
+	Store(memoryBuffer, address, ToUint8(value));
+	Store(memoryBuffer, address + NewInt(b, 1), ToUint8(value >> NewUint(b, 8)));
+	Store(memoryBuffer, address + NewInt(b, 2), ToUint8(value >> NewUint(b, 16)));
+}
+
 void CMemoryUtils::Memory_Write16(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUintValue memoryBuffer, Nuanceur::CIntValue address, Nuanceur::CUintValue value)
 {
 	auto wordAddress = address / NewInt(b, 4);
@@ -74,6 +81,12 @@ void CMemoryUtils::Memory_Write16(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayU
 	AtomicOr(memoryBuffer, wordAddress, valueWord);
 }
 
+void CMemoryUtils::Memory_Write16(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUint8Value memoryBuffer, Nuanceur::CIntValue address, Nuanceur::CUintValue value)
+{
+	Store(memoryBuffer, address, ToUint8(value));
+	Store(memoryBuffer, address, ToUint8(value >> NewUint(b, 8)));
+}
+
 void CMemoryUtils::Memory_Write8(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUintValue memoryBuffer, Nuanceur::CIntValue address, Nuanceur::CUintValue value)
 {
 	auto wordAddress = address / NewInt(b, 4);
@@ -82,6 +95,11 @@ void CMemoryUtils::Memory_Write8(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUi
 	auto valueWord = value << shiftAmount;
 	AtomicAnd(memoryBuffer, wordAddress, mask);
 	AtomicOr(memoryBuffer, wordAddress, valueWord);
+}
+
+void CMemoryUtils::Memory_Write8(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUint8Value memoryBuffer, Nuanceur::CIntValue address, Nuanceur::CUintValue value)
+{
+	Store(memoryBuffer, address, ToUint8(value));
 }
 
 void CMemoryUtils::Memory_Write4(Nuanceur::CShaderBuilder& b, Nuanceur::CArrayUintValue memoryBuffer, Nuanceur::CIntValue nibAddress, Nuanceur::CUintValue value)