--- a/Source/gs/GSH_Vulkan/GSH_VulkanMemoryUtils.h
+++ b/Source/gs/GSH_Vulkan/GSH_VulkanMemoryUtils.h
@@ -31,9 +31,9 @@ namespace GSH_Vulkan
 
 		static void Memory_Write32(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUintValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
 		static void Memory_Write24(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUintValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
-		static void Memory_Write24(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUcharValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
+		static void Memory_Write24(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUcharValue, Nuanceur::CArrayUshortValue,Nuanceur::CIntValue, Nuanceur::CUintValue);
 		static void Memory_Write16(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUintValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
-		static void Memory_Write16(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUcharValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
+		static void Memory_Write16(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUshortValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
 		static void Memory_Write8(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUintValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
 		static void Memory_Write8(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUcharValue, Nuanceur::CIntValue, Nuanceur::CUintValue);
 		static void Memory_Write4(Nuanceur::CShaderBuilder&, Nuanceur::CArrayUintValue, Nuanceur::CIntValue, Nuanceur::CUintValue);