--- a/Source/gs/GSH_Vulkan/GSH_VulkanTransferLocal.cpp
+++ b/Source/gs/GSH_Vulkan/GSH_VulkanTransferLocal.cpp
@@ -153,6 +153,7 @@ Framework::Vulkan::CShaderModule CTransferLocal::CreateShader(const PIPELINE_CAP
 	{
 		auto inputInvocationId = CInt4Lvalue(b.CreateInputInt(Nuanceur::SEMANTIC_SYSTEM_GIID));
 		auto memoryBuffer = CArrayUintValue(b.CreateUniformArrayUint("memoryBuffer", DESCRIPTOR_LOCATION_MEMORY));
+		auto memoryBuffer8 = CArrayUint8Value(b.CreateUniformArrayUint8("memoryBuffer8", DESCRIPTOR_LOCATION_MEMORY));
 		auto srcSwizzleTable = CImageUint2DValue(b.CreateImage2DUint(DESCRIPTOR_LOCATION_SWIZZLETABLE_SRC));
 		auto dstSwizzleTable = CImageUint2DValue(b.CreateImage2DUint(DESCRIPTOR_LOCATION_SWIZZLETABLE_DST));
 
@@ -235,14 +236,14 @@ Framework::Vulkan::CShaderModule CTransferLocal::CreateShader(const PIPELINE_CAP
 		{
 			auto address = CMemoryUtils::GetPixelAddress<CGsPixelFormats::STORAGEPSMCT16>(
 			    b, dstSwizzleTable, dstBufAddress, dstBufWidth, dstPos);
-			CMemoryUtils::Memory_Write16(b, memoryBuffer, address, pixel);
+			CMemoryUtils::Memory_Write16(b, memoryBuffer8, address, pixel);
 		}
 		break;
 		case CGSHandler::PSMT8:
 		{
 			auto address = CMemoryUtils::GetPixelAddress<CGsPixelFormats::STORAGEPSMT8>(
 			    b, dstSwizzleTable, dstBufAddress, dstBufWidth, dstPos);
-			CMemoryUtils::Memory_Write8(b, memoryBuffer, address, pixel);
+			CMemoryUtils::Memory_Write8(b, memoryBuffer8, address, pixel);
 		}
 		break;
 		case CGSHandler::PSMT4HL: