@ericsonia 2021-02-26
Impresive! I hope in the future vulkan feature available for android version