@rcaridade145 2021-02-26
Are you testing on Nvidia? On Virtua Fighter 4 and Marvel Vs Capcom 2 with the Vulkan renderer on master i have slower performance than OpenGL and lots of on screen artifacts.