@Zer0xFF 2021-02-26
> 
> 
> Are you testing on Nvidia? On Virtua Fighter 4 and Marvel Vs Capcom 2 with the Vulkan renderer on master i have slower performance than OpenGL and lots of on screen artifacts.

Yes, im running nVidia RTX 3080.

This PR isnt a comparsion to OpenGL, its Vulkan to itself, and its purely performance, its not meant to fix any of the current artifact issues, though it shouldnt introduce any artifcats as well (if done correctly, unless some GPUs dont support 8/16 bit operations?).