@rcaridade145 2021-02-26
@Zer0xFF I got that. I wasn't comparing Vulkan to OGL. Just trying to explain that on Vulkan renderer i already have artifacts on some games and wanted to question if those artifacts come from the same "area". I'll open a bug about VF4.