@extherian 2021-03-01
Is there any particular reason why Vulkan isn't supported on Android? I thought it was the same API on all platforms, unlike OpenGL and OpenGL ES which are very different from each other.