@Zer0xFF 2021-03-31
ah nice, I've done quick comparsion tests (FFXII, KH, MMX8, CoN) between that branch CI build and local build with my changes & that branch merged, and it seems that branch matches the perfomance of the changes of this PR, without the complication of introducing 8/16bit writes.