@jpd002 2021-03-31
Though, I think having 8/16bit support is helpful because there's still a lot more to add in the Vulkan renderer (upscaling, mipmap support, etc.). I'm pretty sure that GPU usage is lower with your branch added to mine on the games you've mentioned.