@Zer0xFF 2021-04-01
ah, I didnt check GPU usage, so that'd make sense.

well, this should be mostly done, only issue annoying me now, is `Memory_Write16`, it seems when using 16 bit writes or even 2x8 writes in WriteToDepthbuffer we get some corruption (priobably due to incorrect depth)