@jpd002 2021-04-27
Working great with all the games I've tested. Merging this!

Thanks!