@Zer0xFF 2021-04-27
👍 just to note, this is fine as is, the only problem is have (which isnt part of the PR) is that this doesnt seem to work correctly with `GSH_VulkanDraw` which is a vertex shader, though my quick test (writing to fix memory location) suggest it should work just fine.