@jpd002 2021-04-28
Yeah, I don't see why these 8-bit/16-bit memory accesses functions wouldn't work in draw. We just need to be wary of the `coherent` memory access thing. I think it should be fine if the 8/16-bit buffers are marked as `coherent`, but we never know 😛 (Been discovering so much about GPU synchronization lately).